+++ common/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0"
 kinode_process_lib = { git = "https://github.com/kinode-dao/process_lib", rev = "43afdf4" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
